feat: 아이디찾기 삭제로 인한 공통컴포넌트 수정 및 파일명 변경. 비밀번호 찾기 구현
