+++ src/pages/auth-pages/FindPwPage.tsx
@@ -1,20 +1,37 @@
 import { MouseEvent, useState } from 'react'
+import FindPwForm from 'components/auth/FindPwForm'
+import MessageModal from 'components/common/MessageModal'
 import AuthLayout from 'components/auth/AuthLayout'
-import FindIdPwModal from 'components/auth/FindIdPwModal'
-import FindIdPwForm from 'components/auth/FindIdPwForm'
+import { PAGE_LOGIN } from 'constants/auth/path'
+import { FindPwInputValue, Message } from 'types/authType'
+import { userService } from 'services/auth/userService'
 
 const FindPwPage = () => {
+  const [inputValue, setInputValue] = useState<FindPwInputValue>({ email: '', name: '' })
   const [modal, setModal] = useState<boolean>(false)
+  const [modalMessage, setModalMessage] = useState<Message>({
+    status: 'waiting',
+    message: '',
+  })
 
-  const findPw = (e: MouseEvent<HTMLButtonElement>) => {
-    e.preventDefault()
+  const openModal = ({ status, message, goTo }: Message) => {
     setModal((prev) => !prev)
+    setModalMessage({ status, message, goTo })
+  }
+
+  const findPw = async (e: MouseEvent<HTMLButtonElement>) => {
+    e.preventDefault()
+
+    const { result, message } = await userService.findPw(inputValue)
+    openModal({ status: result, message, goTo: PAGE_LOGIN })
   }
 
   return (
     <AuthLayout>
-      <FindIdPwForm findWhat="password" find={findPw} />
-      {modal && <FindIdPwModal modal={modal} setModal={setModal} />}
+      <FindPwForm inputValue={inputValue} setInputValue={setInputValue} findPw={findPw} />
+      {modal && (
+        <MessageModal modal={modal} setModal={setModal} modalMessage={modalMessage} />
+      )}
     </AuthLayout>
   )
 }
